reexport cleanup
