+++ .github/workflows/CompatHelper.yml
@@ -3,8 +3,8 @@ on:
   push:
     branches:
 #      - main #fails, see CI
-        - dev
-    tags: ['*']
+#        - dev
+#    tags: ['*']
 #  pull_request:
 # schedule:
 #   - cron: 0 1 * ? *
